consolidating helper classes
