+++ src/main/java/com/marginallyclever/convenience/helpers/DrawingHelper.java
@@ -1,8 +1,9 @@
-package com.marginallyclever.convenience;
+package com.marginallyclever.convenience.helpers;
 
 import com.jogamp.opengl.GL2;
 import com.jogamp.opengl.util.texture.Texture;
 import com.jogamp.opengl.util.texture.TextureIO;
+import com.marginallyclever.convenience.FileAccess;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
