+++ src/main/java/com/marginallyclever/convenience/helpers/MathHelper.java
@@ -1,4 +1,4 @@
-package com.marginallyclever.convenience;
+package com.marginallyclever.convenience.helpers;
 
 public class MathHelper {
