+++ src/main/java/com/marginallyclever/convenience/helpers/StringHelper.java
@@ -1,4 +1,4 @@
-package com.marginallyclever.convenience;
+package com.marginallyclever.convenience.helpers;
 
 import java.util.Locale;
